+++ concordium-consensus/src/Concordium/GlobalState/Persistent/BlobStore.hs
@@ -148,6 +148,7 @@ import Concordium.Logger (MonadLogger)
 import Concordium.Types
 import Concordium.Types.Accounts
 import Concordium.Types.Updates
+import Concordium.Types.Transactions
 import Concordium.Wasm
 
 import qualified Concordium.Crypto.SHA256 as H
@@ -1385,6 +1386,7 @@ instance MonadBlobStore m => BlobStorable m BakerPoolRewardDetails
 instance MonadBlobStore m => BlobStorable m DelegatorCapital
 instance MonadBlobStore m => BlobStorable m BakerCapital
 instance MonadBlobStore m => BlobStorable m CapitalDistribution
+instance MonadBlobStore m => BlobStorable m SpecialTransactionOutcome -- use the serialize instance
 
 newtype StoreSerialized a = StoreSerialized { unStoreSerialized :: a }
     deriving newtype (Serialize)